Serve a SavedModel

Serve a TensorFlow SavedModel as a local web api under http://localhost:8089.

serve_savedmodel(model_dir, host = "127.0.0.1", port = 8089,
  daemonized = FALSE, swagger = TRUE)

Arguments

model_dir
The path to the exported model, as a string.

host
Address to use to serve model, as a string.

port
Port to use to serve model, as numeric.

daemonized
Makes 'httpuv' server daemonized so R interactive sessions are not blocked to handle requests. To terminate a daemonized server, call 'httpuv::stopDaemonizedServer()' with the handle returned from this call.

swagger
Should a https://swagger.io landing page be made available?

Examples

# NOT RUN {
library(tensorflow)
sess <- tf$Session()

# (1) Train MNIST model.

# (2) Save model with signature.

model_dir <- "trained"
builder <- tf$saved_model$builder$SavedModelBuilder(model_dir)
builder$add_meta_graph_and_variables(
  sess,
  list(
    tf$python$saved_model$tag_constants$SERVING
  ),
  signature_def_map = list(
    serving_default = tf$saved_model$signature_def_utils$build_signature_def(
      inputs = list(images = tf$saved_model$utils$build_tensor_info(x)),
      outputs = list(scores = tf$saved_model$utils$build_tensor_info(y))
    )
  )
)
builder$save()

# (3) Serve saved model.

serve_savedmodel(model_dir)
# }