Predict using a SavedModel

Runs a prediction over a saved model file, local service or cloudml model.

predict_savedmodel(instances, model = NULL, type = c("export", "cloudml",
  "webapi", "graph"),
  signature_name = tf$saved_model$signature_constants$DEFAULT_SERVING_SIGNATURE_DEF_KEY,
  ...)

Arguments

instances
A list of prediction instances to be passed as input tensors to the service. Even for single predictions, a list with one entry is expected.

model
The model as a local path, REST url, CloudML name or graph object.

type
The type of object performing the prediction. Valid values are cloudml, export, webapi and graph.

signature_name
The named entry point to use in the model for prediction.

...
Additional arguments, currently not in use.